Move methods related to file into FileRelatedActionsHelper
