+++ libs/statemanager/src/main/kotlin/mono/state/onetimeaction/FileRelatedActionsHelper.kt
@@ -4,19 +4,64 @@
 
 package mono.state.onetimeaction
 
+import mono.bitmap.manager.MonoBitmapManager
+import mono.export.ExportShapesHelper
+import mono.shape.clipboard.ShapeClipboardManager
+import mono.shape.serialization.SerializableGroup
+import mono.shape.serialization.ShapeSerializationUtil
 import mono.shape.shape.RootGroup
+import mono.state.FileMediator
 import mono.state.StateHistoryManager
 import mono.state.command.CommandEnvironment
 
 /**
- * A helper class to handle file related one-time actions.
+ * A helper class to handle file-related one-time actions in the application.
+ * This class provides methods such as creating new projects, saving and loading shapes to/from
+ * files, exporting selected shapes to text format, etc.
  */
 internal class FileRelatedActionsHelper(
     private val environment: CommandEnvironment,
-    private val stateHistoryManager: StateHistoryManager
+    private val stateHistoryManager: StateHistoryManager,
+    bitmapManager: MonoBitmapManager,
+    shapeClipboardManager: ShapeClipboardManager
 ) {
+    private val fileMediator: FileMediator = FileMediator()
+    private val exportShapesHelper = ExportShapesHelper(
+        bitmapManager::getBitmap,
+        shapeClipboardManager::setClipboardText
+    )
+
     fun newProject() {
         replaceWorkspace(RootGroup(null)) // passing null to let the ID generated automatically
+    }
+
+    fun saveCurrentShapesToFile() {
+        val serializableRoot = environment.shapeManager.root.toSerializableShape(true)
+        fileMediator.saveFile(ShapeSerializationUtil.toJson(serializableRoot))
+    }
+
+    fun loadShapesFromFile() {
+        fileMediator.openFile { jsonString ->
+            val serializableRoot = ShapeSerializationUtil.fromJson(jsonString) as? SerializableGroup
+            val rootGroup = serializableRoot?.let { RootGroup(it) }
+            replaceWorkspace(rootGroup)
+        }
+    }
+
+    fun exportSelectedShapes(isModalRequired: Boolean) {
+        val selectedShapes = environment.getSelectedShapes()
+        val extractableShapes = when {
+            selectedShapes.isNotEmpty() ->
+                environment.workingParentGroup.items.filter { it in selectedShapes }
+
+            isModalRequired ->
+                listOf(environment.workingParentGroup)
+
+            else ->
+                emptyList()
+        }
+
+        exportShapesHelper.exportText(extractableShapes, isModalRequired)
     }
 
     private fun replaceWorkspace(rootGroup: RootGroup?) {
@@ -26,4 +71,3 @@ internal class FileRelatedActionsHelper(
         }
     }
 }
-
